feat(application-save.yml): 配置Knife4j和优化时间戳转换
- 新增Knife4j配置，无需在每个控制器类上添加@EnableKnife4j注解。
- 通过修改时间戳转换逻辑，解决Knife4j显示问题。
- 调整消息转换器位置，确保正确处理时间戳。

fix(时间戳转换): 修正时间戳转换顺序，提升API文档质量和可用性

通过调整消息转换器的顺序，修正了导致Knife4j显示时间戳问题的错误配置。此更改确保时间戳现在能够在API文档中正确显示，提升了文档质量和可用性。

refactor(公共字段填充): 简化和重构公共字段填充逻辑

移除了一些冗余的字段填充代码，优化了实体类的创建和更新时间戳逻辑。这些改动有助于减少代码复杂性，提高可维护性。

feat(用户地址实体):增加用户地址实体类支持

新增UserAddress类，用于存储用户的IP地址信息，包含创建和更新时间戳字段。这为用户地址管理提供了必要的数据结构支持。feat(委托实体): 实现Entrust类以支持委托功能

新增Entrust类，用于存储用户发布的委托信息，包括委托需求、价格、状态和时间戳。这是实现委托功能的基础。

feat(公共基础配置): 引入Knife4j增强配置，提升API文档质量引入Knife4j增强配置，以提升API文档的可读性和可用性。现在可以通过访问/doc.html查看增强的API文档。

refactor(消息转换器): 调整消息转换器，解决Knife4j显示问题

调整消息转换器的顺序，解决了Knife4j无法正确显示文档的问题。此更改确保时间格式正确转换，增强了API文档的质量。

feat(配置): 配置Knife4j增强和时间戳转换优化

- 配置Knife4

+++ public_resource/src/main/java/com/spz/public_resouce/common/BaseContext.java
@@ -4,13 +4,13 @@ package com.spz.public_resouce.common;
  * 基于ThreadLocal封装工具类，用于保存和获取当前登录用户id
  * */
 public class BaseContext {
-    private static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    private static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(int id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static int getCurrentId() {
         return threadLocal.get();
     }
